requirements.txt
pipecat-ai[assemblyai,cerebras,rime,silero,daily,webrtc]
python-dotenv
fastapi
uvicorn
openai
